+++ port/win32_include/omrsignal.h
@@ -26,7 +26,7 @@
 #include "omrcomp.h"
 
 typedef struct J9Win32SignalInfo {
-	uint32_t type;
+	uint64_t type;
 	void *handlerAddress;
 	void *handlerAddress2;
 	struct _EXCEPTION_RECORD *ExceptionRecord;
