+++ port/zos390/omrsignal_context.h
@@ -44,7 +44,7 @@ typedef struct J9PlatformSignalInfo {
 } J9PlatformSignalInfo;
 
 typedef struct OMRUnixSignalInfo {
-	uint32_t portLibrarySignalType;
+	uint64_t portLibrarySignalType;
 	void *handlerAddress;
 	void *handlerAddress2;
 	siginfo_t *sigInfo;
